requirements.txt
streamlit==1.21.0
pandas==1.5.3
numpy==1.24.3
PyPDF2==3.0.1
pdfplumber==0.9.0
python-docx==0.8.11
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==9.5.0
spacy==3.5.3
dateparser==1.1.8
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.0
matplotlib==3.7.1
fpdf==1.7.2
xlsxwriter==3.1.0